The delete user in the setting is done
