+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/Settings/SettingScreen.kt
@@ -5,6 +5,10 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import com.example.habittrackerapp.LocalNavController
 import com.example.habittrackerapp.data
 
@@ -12,6 +16,7 @@ import com.example.habittrackerapp.data
 @Composable
 fun SettingScreen(){
     val navController = LocalNavController.current
+    var popupControl by rememberSaveable { mutableStateOf(false) }
 
     Column {
         Button(onClick = { navController.navigate(Routes.Profile.route) }) {
@@ -20,9 +25,8 @@ fun SettingScreen(){
         Button(onClick = { /*TODO*/ }) {
             Text(text = "signOut")
         }
-        Button(onClick = { /*TODO*/ }) {
-            Text(text = "Delete Account")
-        }
+
+        DeleteUser()
 
         Button(onClick = { /*TODO*/ }) {
             Text(text = "Change Theme")
